Bug 706594: Postscript - Type 1 glyph name equivalence "mapping"
When we load a Type 1 font from a font file (i.e. not an embedded font), we use
the Adobe Glyph List to find if a glyph name has other names (based on the
Unicode code point).

For example, "/ocyrillic" is code point 0x43e, which also commonly maps to the
name "/afii10080".

Previously, use used "forall" to interate through the CharStrings dictionary,
but that causes two problems. Firstly, and most importantly, when we write new
entries to that dictionary, if the dictionary has to be extended, it ends up
messing with the "forall" indexing. Secondly, it means we do work than necessary
because we potentially seek out equivalents for names we've just added.

To improve this, populate an array with the original names from the CharStrings
dictionary, and iterate through that - thus the changing contents of the
dictionary doesn't matter.