GhostPDF - set the default for the key length when /V is 1
Bug #706668 "regression: new PDF engine can't handle a broken PDF file, falsely reports password protected"

Although the file is damaged (in fact it appears to be 2 PDF files,
concatenated) that's not the problem. The second file is the one which
we process, and that file is encrypted. We simply were not setting the
encryption key length for Revision 3 and we were only reading the
actual key /Length if /V is present and has the value 2 or 3.

This is basically because PDF encryption is a mess, and simply an
oversight. Fix it here.