Bug 706643: Fix fill_stroke transparency handling.
The code to handle fill_stroke in the presence of transparency
was being confused by the scaling applied to clipping regions.

When we fill_stroke, if transparency is present, then we need to
push a group. The bounds for this group are calculated based upon
the bounds for the path in question, expanded by an appropriate
amount for the current stroke strate, and then clipped by the
clipping path.

In the presence of alphabits, the path is scaled up, and so the
bbox for it needs to be scaled down. The clipping path is similarly
scaled, so this too needs to be scaled down.