Bug 706683: Fill+Stroke: Keep clist in sync with transparency changes.
Currently, it looks like we are hitting a problem with the fix for
bug 703324.

When that fix resets the fillstrokealpha value after the stroke
it does so by fiddling directly in the pgs. In the clist writing
case, this means that the pgs changes without the change being
considered for putting into the reader. This can cause the reader
and writer to get out of sync.

The fix, implemented here, is to call gs_update_trans_marking_params
to ensure that the change happens via the compositor mechanism
first. This keeps reader and writer in sync.