GhostPDF - fix buffer overrun reading corrupted CFF font
OSS-fuzz #58405

There is a guard to prevent buffer overruns, but it wasn't taking the
NULL terminator into account. In addition, I think it is possible for
the required number of bytes to be 4, not 3, if the byte pointed to
is 0xCC (resulting in 'E-', 'E-' being generated) and then still
potentially requiring a NULL terminator for a total of 5 bytes.

Change the 3-byte minimum space requirement to 5.