gpdl: Support Multipage TIFFs.
This supports the common form of Multipage TIFFs, whereby a single
TIFF has multiple IFDs.

We do not yet support the (much rarer) form of multipage TIFFs
whereby we have a single IFD with sub-ifd's in it.

Those are supposedly used for strange things like multiple subsamplings
of the same image, so are probably not actually what we want anyway.