mswinpr2 device - permit use of AlphaBits
Bug #693636 "Antialiasing options are ignored in windows printer driver"

This commit processes the Graphics and Text AlphaBits parameters and
stores them. When we later set the actual colour depth of the device
we set the device color_info structure anti-alias bits from the
requested AlphaBits, taking the actual colour depth of the device into
account for the purposes of setting the maximum.