mswinpr2 device - honour User settings with default printer
Bug #693637 "Some of user options are not working in windows printer driver when option QueryUser set to 3."

Some (undocumented!) User settings which can only be altered by
sending PostScript were not being applied when the QueryUser
setting was 3 (use system default printer). This was despite
the Copies being set in this case, so this was clearly an oversight.