Rework 'is_planar' flag in device header.
Replace 'is_planar' with 'num_planar_planes'. 0 still means
not planar, and non-zero means planar with that number of
planes.

This is intended to break the assumption, carried throughout
our planar code, that num_planes == color_info.num_components.

For an rgbtags device, we need color_info.num_components to
remain as 3, where num_planes is 4.

bitrgbtags works with num_components=3 because it is chunky.
A hypothetical psdrgbtags would have problems though, because
either we'd need to have color_info.num_components = 4 (which
would trip various parts of the graphics system into thinking
we were working with CMYK), or we'd not get storage defined
for the tags.