Tiffsep - Close separation files when ReopenPerPage is true
Bug #694859 "tiffsep leaves handle open on main generated composite file when called from within code in .net"

The behaviour in the bug report is actually expected for a fixed single
filename (eg output.tif), but while investigating it I ran across the
undocumented ReopenPerPage parameter. If this is true then gdevprn will
close the output file after every page, meaning it has to be reopened
at the start of the next page, instead of just retaining the old file.

Using this solved the problem for the composite file, but left all the
separation files still open. This commit honours the ReopenPerPage
parameter for the separation files too.