PCL-XL device - alter text to vectors instead of bitmaps
Bug 696224 "Invalid font colors produced when using freetype renderer"

To summarise the lengthy thread; The PXL device doesn't write fonts.
text was emitted as bitmaps and done in a way which basically involves
PCL RasterOPs.

These are incompatible with the pdfwrite device and also, according to
the thread, with some PCL clone printers.

Moving to using fonts would be the best solution, but this would be a
large project. If it were only pdfwrite, then I would simply close the
bug as we know RasterOPs don't work with pdfwrite. However the fact that
it also apparently fails on some printers *and* doesn't work properly
with some of our own devices I think it needs fixing.

Two solutions were suggested and Henry wanted to know which provided
'better' results. I've tested both and neither is great. Both result in
larger file sizes in general. The file attached to the bug report
actually produces a smaller file with the route I've chosen but that's
anomalous, in general files with text will be larger. A 29 page test
file, mostly text, is 7 times larger with this commit. However testing
Henry's proposed solution resulted in a file 25 times larger.

Here we simply disable the glyph cache, so we never create a bitmap and
write the text as vectors. This results in the correct colours being
produced by pdfwrite and I hope by the clone PCL printers as well.