GhostXPS & GhostPCL - fix a memory leak
Bug #700569 "memory leak with pdfwrite"

The pdfwrite device copies glyph names, and expects (because of its
heritage with the PostScript interpreter) that the glyph names will
persist until the PDF file is written.

In order to fulfill this assumption the XPS and PCL interpreters
allocate new strings to return the glyph name. However prior to this
commit they did no track these strings, leading to memory leaks.

In this commit I've added a 'names' array to the font 'wrapper' used
by the interpreters, as well as a couple of integers to track how many
names have been allocated and (for performance reasons) a maximum, so
that we can allocate new table sin chunks as needed.

When we allocate a name for a font we store the pointer in the array,
when we free the font wrapper we free these strings.