xpswrite - fix blank page emission
Bug 695643 "Page missing with xpswrite device"

The xpswrite device defers writing the page opening boilerplate until
some marking operation takes place. The problem was that if the page did
not contain any marking operations, we never emitted the page opening
code (though we did write the closing code!)

This commit just checks, on outputting a page, if the opening code has
been written and emits it if it hasn't.