PS interpreter - handle gs_scan_string_token() re-allocating the op stack
Bug #706737 "Heap buffer overflow write of size 8 in gs_scan_token (iscan.c:669)"
Also Bug #706741

The problem is that we save the operand stack pointer, then call
gs_scan_string_token(), then call the 'push' macro which sets the
operand stack pointer back to the saved pointer, and adds the required
number of objects.

However, gs_scan_string_token() can encounter a stack overflow and in
that case it tries to grow the operand stack. That reallocates the stack
and invalidates our saved pointer. Deal with that by refreshing the
pointer just in case the stack has moved.

The bug only encountered this with a token operator, but grepping the
code shows two other places with the same problem. Fix those too.

Our special thanks to Alan Goodman for the fuzz testing which uncovered
this bug.