Fix commit 78da18e1d130468a171ef6aca58b2ec23841332b
Silly mistake on my part. We use font->font->memory to find the
allocator used to free the stored names, but we were using it *after*
we had freed font->font.....

Just move the code so that we use it before we free it.

Came up in the overnight debug-Z@ tests, surprisingly on only one file.