PCL interpreter - Fix disk-based macros
bug #706747 "Calling disk-based macro form exits with error code -1"

Commit cea2da4aae1edbf6aef128f44962cfe8baa5c728 added support for fonts
on disk, as part of that the code was reformatted and "better error
handling for alphanumeric key support" added.

Unfortunately, that broke disk-based macros, and I suspect didn't work
with fonts, if the macro (or font) was called for more than once.

The reason is that we try to associate the string ID of the macro with
the 'current macro' ID. That isn't a problem and works fine. However
the *second* time we try to do that we have already associated the
string ID with the current macro ID, so the pl_dict_put_synonym() call
exits with an error. The code prior to the commit silently ignored the
error, the current code returns it. Which is what causes the failure.

I don't know enough to try and tackle the basic cause, but simply
restoring the old behaviour of ignoring the error works for the file
here and the file for bug #697900 (the original bug report for
disk-based fonts not working).