pdfi: Manage text state for pattern PaintProcs
When embark on executing a pattern PaintProc, we want to start with a blank
text state. So store, reset and restore the text state before and after
calling the PaintProc.

Avoids confusion when a Type 3 CharProc draws using a Type 1 pattern.

Comes from oss-fuzz 59023, but it's not clear at the moment if it's the root
of the problem oss-fuzz found.