PCL interpreter - fix command scanner
Bug 695047 "'ESC * r <code> U' for setting Simple CoLor Space to -4 (KCMY) not implemented"

This commit doesn't actually fix the bug as described. The problem it
fixes is, however, exhibited by the file in this bug report.

The problem is that we encounter a 'transfer raster row' operation, but
do not have sufficient bytes in the buffer to satisfy the number of
bytes required to be transferred.

This causes us to return to copy the data we do have and return to the
input loop for more data. Once we have retrieved the data we should
(obviously) execute the transfer raster row operation. What actually
was happening was that the 'continue' would cause us to set the
function 'cdefn' to NULL, and the fact that we had set the scan type to
'scanning_parameter' would mean we went back to looking for parameter
data. Which we'd find because we'd be pointing at the *next* transfer
raster row operation.

So we'd read the parameter, determine the number of bytes by
incorporating it into the existing number of bytes (from the previous
operation). That would then lead to us trying to read far too many bytes
for the transfer raster row. After that we'd try to parse the data as
PCL, but would fail, and so we'd fall back to handling it as text.

Removing the continue means we carry on and process the interrupted
operation (transfer raster row), and *then* go on to scanning for
parameters.

With this (and the previous commit to address mode 9 decompression) the
tiger decompresses recognisably. It is not 100% correct, as noted in
bug 694990.