PCL interpreter - fix mode 9 decompression
BUG #694990 "mode 9 uncompress simply does not work correctly."

With this change (in addition to what's already committed), the usual
tiger shows with:
https://bugs.ghostscript.com/attachment.cgi?id=10631

But now with wrong color.

I believe this is the missing part of the "better patch" I wrote in:
https://bugs.ghostscript.com/show_bug.cgi?id=694990#c2

The comment appears to be confusing/wrong; given now that we have an
implementation which works on a test file, it is probably best to remove
the comment.