PS interpreter - remove GS-specific binary object sequence extension
Bug #706743 "Segfault on unknown address in dict_find (idict.c:293)"

This is a subtle problem caused partially by the way Ghostscript
represents dictionaries.

The BOS in this file executes a sequence :

BOS header
0x81 - BOS, low order byte first IEEE standard real format
0x02 - normal header, 2 objects
0x3030 - length 12336 bytes

Object 1
0x0F - literal, type = GS extension, dictionary
0x00 - unused, must be 0
0x0000 - length of object (0 bytes)
0x000000 - value, usage ?

Object 2
0x06 - literal, type = immediately evaluated name
0x00 - unused must be 0
0xFFFF - -1 = system name index
0x00000165 = stack

The BOS code stores a dictionary as a 'mixed' array type. In order to
fixup objects it then processes the array contents, starting with the
names, and when all the names are processed by looking for any
'mixedarray' types, which it takes to be dictionaries and processes.

The immediately evaluated name causes the second object to be replaced;
because the 'stack' operator is actually implemented by Ghostscript as
a procedure, this means that the second object is replaced by (tada!) a
mixed array.....

The BOS code then tries to handle that mixed array as a dictionary,
and that causes a crash.

The code has a comment stating that names need to be processed *before*
dictionaries, so there is no way, without considerable rewriting, of
resolving this. However; the use of dictionaries in binary object
sequences is totally non-standard. BOS are extremely rare in PostScript
anyway, and obviously the only practical use of this extension would be
by Ghostscript itself (such as processing startup code).

Since we don't use this extension, and it's non-standard, the simplest
solution is to just remove the extension.

Our special thanks for Alan Goodman for the fuzz testing which
uncovered this bug.