Bug 706744: Immediate evaluation of a currentfile reference
Doing "/f currentfile def //f" resulted in the working ref (token) being pointed
at by the "object to be interpreted" ref (IREF) since both were the
"currentfile". But this meant that the next time around the interpreter loop
IREF could remain pointing at token when it was no longer an object suitable
for interpreting.

The above Postscript snippet is the only case where this can happen.

So, in event it does happen, find the currentfile ref stored on the execution
stack, and point IREF at that. Although both that and "token" will reference
the same stream object, there are different ref objects, this token can be
overwritten when the next interpreted object, without corrupting the interpreter
loop access to the current file object.

Our special thanks for Alan Goodman for the fuzz testing which
uncovered this bug.