Fix timeout seen in cluster testing.
  gs -o out%d.psd -Z@ -dMaxBitmap=10000 -sDEVICE=psdcmyk16 -r300 -Z:
     -sDEFAULTPAPERSIZE=letter -dNOPAUSE -dBATCH -K2000000 -dClusterJob
     -dJOBSERVER tests_private/pdf/PDFIA1.7_SUBSET/CATX6508.pdf

Another case where subdivide_patch_fill was getting into an
infinite loop. But for a different reason this time.

When entering the loop, we were checking for the patch being
completely off the page. Then we check for edges of the patch
being off the left or right of the page, and truncate the
patch appopriately. Then we try to do the same for the top
and bottom of the page.

BUT, after having truncated because of left/right clipping,
we are no longer guaranteed that the entire patch isn't off
the top/bottom, hence our preconditions fail.

The easiest fix for this is to split the 'completely off the
page' test into two. X first, then Y later.