PDF interpreter - avoid potential hang reading data
Bug #706757 "Potential hang reading PDF"

The crazy way the stream code works with ASCII85Decode; if we reach the
end of data, skip that, and then try to read more data, and the string
data is followed by a CR/LF, then we would try to read those last two
bytes.

Because we are already at EOD, we wouldn't actually read them, but
would move the buffer pointer on. So seofp() would decide that we are
not at the end of the data. This would then result in an infinite loop
because we would read no bytes, but the seofp() test would always fail
so we'd try again to read bytes.

By moving the sbufskip() in the loop we can avoid this situation.