Bug 706755: Fix gpdl SEGV with fuzzed JPEG file.
The fuzzed jpeg in question is so broken, that we would never
ordinarily have used the JPEG handler. If we force it to, then
it can fail at an earlier stage than anticipated and attempt
to use an unintialised pointer to a jmpbuf.

Fix this by aligning the jmpbuf earlier.