xpswrite - set the winding rule on filled paths
Bug #695641 "Filling differences"

This turns out to be because the PDF file has a filled self-intersecting
path where the winding rule makes a difference to the 'insideness'.

The xpswrite device simply was not writing the winding rule at all!

Not only that, but the default for XPS is Even-Odd while the default for
PDF is non-zero. Not surprisingly this causes progressions in quite a
number of XPS output files from PDF input.