Bug 706752: Fix save/restore interaction with full clumps.
We have an optimisation that reduces the time spent in garbage collection, so
that, when we create a save state we remove the "has_refs" flag from the
allocator clumps so that we don't spend time scanning for refs that will not,
and cannot change. To function properly, the optimisation requires that, when
we do a restore, the original "outer clump" is returned to being current, and
thus the has_refs flag is returned to the correct value for the active clump.

The problem is, there are two (related) circumstances under which we cannot
rely on that. One is single object clumps (c_alone == true): so objects
allocated as "immovable", or objects of sufficient size that they warrant an
entire clump of their own. Second is a regular clump that doesn't have space
for any more allocations.

In neither of these cases do we create an inner clump as part of the save op,
because we know that there cannot be any changes to that clump that we might
have to undo (at the allocator level). Thus restore does not put the has_refs
flag back again, so we don't know we have refs that need remarked as "new".

As a result, if we change the contents of a dictionary whose kay/value arrays
were allocated in such a clump, we can wrongly create a change record for that
change, which will refer to an object which can be freed (as it is exists only
in the current save level). When we then try to access that change record, we
can potentially crash, accessing freed memory.

In those cases, do not apply that optimisation.

Our special thanks for Alan Goodman for the fuzz testing which
uncovered this bug.