Subclassing - better handling of finalization
This arises from bug #706759 because, due to the actual fault in that
bug, we triggered an ASAN error writing to freed memory.

The changes due to commit 510bd436dec7c52cdf7222d3daa4184392453464 and
b609c25f7f881bea4b10c8e1b9b950b635437320 to resolve the way that the
device class finalise works with reference counting left a problem in
place; we were accessing a device after it had been freed.

In fact, as noted in the comment in this commit, none of the fiddling
with pointers is actually necessary; if we are finalising a subclassing
device it can only be because the interpreter/graphics library has
finished with it. Since the subclassing device is pretending to be some
real device, the one at the end of the chain, that means the entire
chain should be freed.

Due to a fault in error handling with stringwidth the final device was
counted up too many times, resulting in it not being freed when it
should have been. The next commit will fix that problem. In this commit
we remove the pointer fiddling, but as we finalise each subclassing
device in the chain we check to see if the child has a reference count
other than 1, if it does we count it down to 1, and in debug mode we
write a warning.