PS interpreter - clean up properly on error in stringwidth
Bug #706759 "Heap use after free write of size 8 in default_subclass_finalize (gdevsclass.c:895)"

The problem here is that we get a stack overflow after calling
gs_stringwidth_begin(). That creates a text enumerator, and one of the
effects of that is to count up the device. because we were never freeing
the enumerator, we tried to exit with the device still counted up too
many times, which (see previous commit) caused us to try and access
free memory (in subclassing).

If we get an error after we have successfully completed
gs_stringwidth_begin() then we need to count down the enumerator before
returning.