Bug 706766: On VMerror, explicitly free a change object
Allocating an extremely large array, we create the save object before attempting
to allocate the array. If the array allocation fails, we have an uniniitialised
"unattached" save object, which causes gc to crash later on.

Just free the change before returning the error

Our thanks for Alan Goodman for the fuzz testing which uncovered this bug.