Bug 706765: Don't try to validate a path that's a NULL pointer
A zero length Postscript string does not allocate a string body, just sets it to
NULL. So using a zero length string as the scratch string parameter to
filenameforall results in a NULL pointer being passed into the file enumeration
code, which handles it correctly. But the file permission vaidation code did not
account for that possibility.

In the filenameforall case, we want this to trigger a rangecheck error, but not
in other cases. So, handle that part in the file enumeration code, but also add
an extra check for safety in the validation code.

Thanks for Alan Goodman for the fuzz testing which
uncovered this bug.