xpswrite device - handle clipping
Bug #695646 "Rectangles obscure text"

The problem is that the xpswrite device didn't handle any kind of clip.
The only use of a clip was when drawing with an image brush where the
clip was repurposed into a path for the image to fill.

This commit addresses that, its complicated and probably needs more work
but it fixes the problem in this report, and a number of other PDF files
without causing any regressions.

XPS does not handle clips (or indeed any 'graphics state') the same way
as PostScript and PDF. Clips are either a property of an drawing opeation
(eg stroke or fill) or a property of the canvas. I suspect we should be
dealing with clip by creating sub-canvases when the graphics state
changes, and closing the canvas on a restore, but that would be quite
complicated (see pdfwrite) to implement.

There is another similar bug report, 695647, which is partially
improved with this commit but where problems remain. That bug may well
require a more thorough treatment of clipping.