RAM file system - prevent negative file position
Bug 706773 "OOB write on %ram% files with negative offset"

By carefully constructing in-memory buffers, and then using a
negative file offset, it is possible to access atribtrary
memory using file operations, which is exploitable.

In fact it should not be possible to get into this state, as the
PostScript setfileposition operator is supposed to error on
negative file offsets, I'll address that in a later commit.
It seems best to me to fix this in the underlying code as well.

This commit adds a missed check for the position being negative
as well as a very useful overflow check.

The comments here are from me, since the diff Zdenek supplied
did not have any commit log, any errors or omissions are
therefore my fault.