Bug 706770: Fix reference counting when evaluating transfer functions
Internal transfer function objects are one of the cases where (with Postscript)
we use both reference counting and garbage collection to manage the life spans
of object instances.

During evaluation we store a counted reference to the map on the exec stack,
which we decremented the ref count, before popping the objects off the stack
on successful completion of the evaluation.

In this case, we encounter a typecheck error part way through the evaluation,
meaning we'd have decremented the reference count, but *not* popped the
reference off the exec stack.

To maintain proper Postscript error handling, we don't want to pop the objects
in the event on an error, but we also don't want to retain a reference no
longer reference counted. So, in that event, make that one entry on the stack
null (since it's an internal object, that won't adversely affect error
reporting).

Thanks for Alan Goodman for the fuzz testing which uncovered this bug.