Bug 706782: Fix SEGV with tags device.
When opening the pattern accumulator device for a tags setup
ensure we open it with the correct number of planes, not just
the number of colors.

Cluster testing this shows up various problems to do with the
pdf14 devices changing their color setup without changing
the number of planar planes. Fixes for that are included here.