Fix detection of nulldevice
Since we revised how device procs get filled in, the detection of the nulldevice
has been faulty.

It relies on checking that the fill_path proc in the candidate device is the
same as the fill_path in the nulldevice prototype. But proc pointers in
device prototypes are now null, and they are filled in during device
creation by the "initialize_device_procs" method.

Fix the code to deal with this new devices world.