graphics library - prevent clip rectangle list disappearing while in use
Bug #706771 "Segfault on unknown address in clip_enumerate_rest (gxclip.c:271)"

After considerable simplification the binary token PostScript program
can be reduced to:

  0 0 5 0 48 arcn eoclip
  5 48 16 matrix
  {grestore (0)}
  image

The problem is that the image operator creates a 'clipper' device which
stores a pointer to the clip rectangle list (scan converted path) but,
crucially, does not increment its reference count, even though the
clip_rect_list is reference counted, it just uses the list member
directly without increasing the count.

We then execute a grestore as part of the procedural data source for the
image data, which restores away the clip path. Then we try to draw the
image through the clipper device, which tries to use the clip rectangle
list, which has disappeared.

This whole area is pretty shoddy and I assume it is more or less
untouched since the days before Ghostscript had reference counting. In
my opinion the right way to solve this is to increment the reference
count of the clip rectangle list when we take a reference, and decrement
it when we release it.

That means that we should also have a finalize method for the clipper
device, so that it can do the release when the device is freed.

In order to do that I had to move the definition of gx_clip_rect_list
because otherwise we had two mutually dependent header files, which just
wouldn't compile.

Finally, in gx_image1_end_image() instead of freeing the clipper device
directly, count down it's reference count so that it's finalize method
will correctly release the reference to the clip rectangle list.